feat: Added notification section for quality control expiration. Respective tests added. Re-organization of the test_views.py into smaller python files by component.
