+++ templates/dashboard_cliente.html
@@ -155,6 +155,41 @@
                     </div>
                     {% endif %}
 
+                    {% if equipos_cc_por_vencer %}
+                    <div class="notifications-section mt-4">
+                        <h4 class="mb-3"><i class="fas fa-bell me-2"></i>Notificaciones: Controles de Calidad Próximos a Vencer</h4>
+                        <div class="table-responsive">
+                            <table class="table table-hover table-sm">
+                                <thead class="table-light">
+                                    <tr>
+                                        <th>Nombre</th>
+                                        <th>Marca</th>
+                                        <th>Modelo</th>
+                                        <th>Serial</th>
+                                        <th>Fecha Vencimiento</th>
+                                    </tr>
+                                </thead>
+                                <tbody>
+                                    {% for equipo in equipos_cc_por_vencer %}
+                                    <tr>
+                                        <td>{{ equipo.nombre }}</td>
+                                        <td>{{ equipo.marca }}</td>
+                                        <td>{{ equipo.modelo }}</td>
+                                        <td>{{ equipo.serial }}</td>
+                                        <td>{{ equipo.fecha_vencimiento_control_calidad|date:"d/m/Y" }}</td>
+                                    </tr>
+                                    {% endfor %}
+                                </tbody>
+                            </table>
+                        </div>
+                    </div>
+                    {% else %}
+                    <div class="notifications-section mt-4 text-center">
+                         <h4 class="mb-3"><i class="fas fa-bell me-2"></i>Notificaciones</h4>
+                        <p>¡Felicidades, no hay equipos con controles de calidad próximos a vencer!</p>
+                    </div>
+                    {% endif %}
+
                     {% if procesos_activos_cliente %}
                     <div class="notifications-section mt-4">
                         <h4 class="mb-3"><i class="fas fa-tasks me-2"></i>Mis Procesos Activos</h4>
